another round angles to int bug (this time in PF_vectoyaw)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4511 d7cf8633-e32d-0410-b094-e92efae38249

+++ pr_cmds.c
@@ -522,7 +522,7 @@ void PF_vectoyaw (void)
 		yaw = 0;
 	else
 	{
-		yaw = (int) (atan2(value1[1], value1[0]) * 180 / M_PI);
+		yaw = (atan2(value1[1], value1[0]) * 180 / M_PI);
 		if (yaw < 0)
 			yaw += 360;
 	}
